requirements.txt
psutil==5.4.5
kafka-python==1.4.2